hng_tasks/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==4.2.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==20.1.0
inflection==0.5.1
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
swagger_ui_bundle==1.1.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
